Checkpoint 1

+++ app/src/main/java/com/canque/buddly/models/Budget.kt
@@ -3,6 +3,6 @@ package com.canque.buddly.models
 data class Budget(
     val id: String,
     var name: String,
-    var amount: Float,
+    var amount: Double,
     var category: Category,
 )
